Refactor game organization
